feat: typescript 2
